ast: Take differing axis scales into account when finding length of grid curves
This can affect where interior axis lables are placed. Nothing should
change for grf modules that have always used equal axis scaling.

+++ libraries/ast/plot.h
@@ -671,8 +671,8 @@ typedef struct AstPlotGlobals {
    char 	StripEscapes_Buff[ AST__PLOT_STRIPESCAPES_BUFF_LEN + 1 ];
    double 	Grf_chv_t;
    double 	Grf_chh_t;
-   double 	Grf_alpha_t;
-   double 	Grf_beta_t;
+   float 	Grf_alpha_t;
+   float 	Grf_beta_t;
 } AstPlotGlobals;
 
 #endif
